+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add Roles random table.
 - Tweak various random description tables to be third-person.
 - Tweak foe and vehicle sheet top spacing.
+- Add solarized color schemes.
 
 # 1.1.0